refactor: extract tab runtime factory
