+++ src/features/chat/tabs/Tab.ts
@@ -65,12 +65,8 @@ import {
 } from '../execution/ChatExecutionCoordinator';
 import { MessageRenderer } from '../rendering/MessageRenderer';
 import { cleanupThinkingBlock } from '../rendering/ThinkingBlockRenderer';
-import { createWelcomeElement } from '../rendering/WelcomeRenderer';
 import { findRewindContext } from '../rewind';
 import { BangBashService } from '../services/BangBashService';
-import { SubagentManager } from '../services/SubagentManager';
-import { ChatState } from '../state/ChatState';
-import type { TabAttention } from '../state/types';
 import { BangBashModeManager as BangBashModeManagerClass } from '../ui/BangBashModeManager';
 import { ComposerContextTray } from '../ui/ComposerContextTray';
 import { FileContextManager } from '../ui/FileContext';
@@ -85,17 +81,15 @@ import { recalculateUsageForModel } from '../utils/usageInfo';
 import { getTabProviderId } from './providerResolution';
 import { TabModelSelectionCoordinator } from './TabModelSelectionCoordinator';
 import { TabRuntimeCleanup } from './TabRuntimeCleanup';
-import { TabSession } from './TabSession';
+import { createTabRuntime } from './TabRuntimeFactory';
 import type {
   ProviderCatalogInfo,
   ProviderCatalogResolver,
+  TabCreateOptions,
   TabData,
-  TabDOMElements,
-  TabId,
   TabManagerViewHost,
   TabProviderContext,
 } from './types';
-import { generateTabId } from './types';
 
 type TabProviderSettings = Record<string, unknown> & {
   model: string;
@@ -139,23 +133,6 @@ export function getBlankTabModelOptions(
     return uiConfig.getModelOptions(settings)
       .map(model => ({ ...model, group, providerIcon }));
   });
-}
-
-export interface TabCreateOptions {
-  plugin: FeatureHost;
-
-  containerEl: HTMLElement;
-  conversation?: Conversation;
-  tabId?: TabId;
-  /** Initial draft model for an unbound tab. */
-  draftModel?: string | null;
-  lifecycleState?: Extract<TabData['lifecycleState'], 'provisional' | 'cold'>;
-  onStreamingChanged?: (isStreaming: boolean) => void;
-  onRewindingChanged?: (isRewinding: boolean) => void;
-  onTitleChanged?: (title: string) => void;
-  onAttentionChanged?: (attention: TabAttention) => void;
-  captureReviewableSettlement?: () => () => void;
-  onConversationIdChanged?: (conversationId: string | null) => void;
 }
 
 export { getTabProviderId } from './providerResolution';
@@ -593,148 +570,13 @@ export function onProviderAvailabilityChanged(tab: TabData, plugin: FeatureHost)
   return tab.draftModel !== previousDraftModel || tab.providerId !== previousProviderId;
 }
 
-/**
- * Creates a new Tab instance with all required state.
- */
+/** Creates a new Tab instance with all required state. */
 export function createTab(options: TabCreateOptions): TabData {
-  const {
-    plugin,
-    containerEl,
-    conversation,
-    tabId,
-    onStreamingChanged,
-    onRewindingChanged,
-    onAttentionChanged,
-    onConversationIdChanged,
-  } = options;
-
-  const id = tabId ?? generateTabId();
-
-  const contentEl = containerEl.createDiv({ cls: 'claudian-tab-content claudian-hidden' });
-
-  const state = new ChatState({
-    onStreamingStateChanged: onStreamingChanged,
-    onRewindingStateChanged: onRewindingChanged,
-    onAttentionChanged: onAttentionChanged,
-    onConversationChanged: onConversationIdChanged,
+  const tab = createTabRuntime(options, {
+    onStreamingStateChanged: updateSendButton,
   });
-
-  // Create subagent manager with no-op callback.
-  // This placeholder is replaced in initializeTabControllers() with the actual
-  // callback that updates the StreamController. We defer the real callback
-  // because StreamController doesn't exist until controllers are initialized.
-  const subagentManager = new SubagentManager(() => {});
-
-  const dom = buildTabDOM(contentEl);
-  state.queueIndicatorEl = dom.queueIndicatorEl;
-
-  const isBound = !!conversation?.id;
-  const restoredDraftModel = typeof options.draftModel === 'string'
-    ? options.draftModel.trim()
-    : '';
-  const newConversationModel = !isBound && !restoredDraftModel
-    ? resolveNewConversationModel(plugin.settings)
-    : null;
-  const draftModel = isBound
-    ? null
-    : (restoredDraftModel || newConversationModel?.model || null);
-  const initialProviderId = conversation?.providerId
-    ?? newConversationModel?.providerId
-    ?? (draftModel
-      ? getEnabledProviderForModel(draftModel, plugin.settings)
-      : DEFAULT_CHAT_PROVIDER_ID);
-  const session = new TabSession({
-    id,
-    lifecycleState: options.lifecycleState ?? 'cold',
-    draftModel,
-    providerId: initialProviderId,
-    conversationId: conversation?.id ?? null,
-  });
-  const tab: TabData = {
-    session,
-    get id() {
-      return session.id;
-    },
-    get lifecycleState() {
-      return session.lifecycleState;
-    },
-    set lifecycleState(value) {
-      session.lifecycleState = value;
-    },
-    hydrationState: isBound ? 'idle' : 'ready',
-    get draftModel() {
-      return session.draftModel;
-    },
-    set draftModel(value) {
-      session.draftModel = value;
-    },
-    get providerId() {
-      return session.providerId;
-    },
-    set providerId(value) {
-      session.providerId = value;
-    },
-    get conversationId() {
-      return session.conversationId;
-    },
-    set conversationId(value) {
-      session.conversationId = value;
-    },
-    get executionCoordinator() {
-      return session.executionCoordinator;
-    },
-    set executionCoordinator(coordinator) {
-      session.setExecutionCoordinator(coordinator);
-    },
-    providerCatalogResolver: null,
-    captureReviewableSettlement: options.captureReviewableSettlement ?? null,
-    state,
-    controllers: {
-      selectionController: null,
-      browserSelectionController: null,
-      canvasSelectionController: null,
-      conversationController: null,
-      streamController: null,
-      inputController: null,
-      navigationController: null,
-    },
-    services: {
-      subagentManager,
-      instructionRefineService: null,
-      titleGenerationService: null,
-    },
-    ui: {
-      contextTray: null,
-      fileContextManager: null,
-      imageContextManager: null,
-      modelSelector: null,
-      modeSelector: null,
-      thinkingBudgetSelector: null,
-      externalContextSelector: null,
-      mcpServerSelector: null,
-      permissionToggle: null,
-      serviceTierToggle: null,
-      slashCommandDropdown: null,
-      instructionModeManager: null,
-      bangBashModeManager: null,
-      contextUsageMeter: null,
-      statusPanel: null,
-      navigationSidebar: null,
-    },
-    dom,
-    renderer: null,
-  };
-
-  state.callbacks = {
-    ...state.callbacks,
-    onStreamingStateChanged: (isStreaming) => {
-      onStreamingChanged?.(isStreaming);
-      updateSendButton(tab);
-    },
-  };
   updateSendButton(tab);
-
-  tab.executionCoordinator = createTabExecutionCoordinator(tab, plugin);
+  tab.executionCoordinator = createTabExecutionCoordinator(tab, options.plugin);
   return tab;
 }
 
@@ -1134,55 +976,6 @@ function normalizeProviderMode(mode: string): string {
   if (mode === 'bypassPermissions' || mode === 'yolo') return 'yolo';
   if (mode === 'plan') return 'plan';
   return 'normal';
-}
-
-/**
- * Builds the DOM structure for a tab.
- */
-function buildTabDOM(contentEl: HTMLElement): TabDOMElements {
-  const messagesWrapperEl = contentEl.createDiv({ cls: 'claudian-messages-wrapper' });
-  const messagesEl = messagesWrapperEl.createDiv({ cls: 'claudian-messages' });
-  const welcomeEl = createWelcomeElement(messagesEl);
-  const statusPanelContainerEl = contentEl.createDiv({ cls: 'claudian-status-panel-container' });
-  const inputComposerEl = contentEl.createDiv({ cls: 'claudian-input-composer' });
-  const inputContainerEl = inputComposerEl.createDiv({ cls: 'claudian-input-container' });
-  const queueIndicatorEl = inputContainerEl.createDiv({ cls: 'claudian-input-queue-row' });
-  const navRowEl = inputContainerEl.createDiv({ cls: 'claudian-input-nav-row' });
-  const inputWrapper = inputContainerEl.createDiv({ cls: 'claudian-input-wrapper' });
-  const contextRowEl = inputWrapper.createDiv({ cls: 'claudian-context-row' });
-  const inputEl = inputWrapper.createEl('textarea', {
-    cls: 'claudian-input',
-    attr: {
-      placeholder: 'Ask to make changes, @mention files, run /commands',
-      rows: '3',
-      dir: 'auto',
-    },
-  });
-  const sendButtonEl = inputWrapper.createEl('button', {
-    cls: 'claudian-input-send-button',
-    attr: {
-      type: 'button',
-      'aria-label': 'Send message',
-      title: 'Send message (enter)',
-    },
-  });
-  setIcon(sendButtonEl, 'send');
-
-  return {
-    contentEl,
-    messagesEl,
-    welcomeEl,
-    statusPanelContainerEl,
-    inputComposerEl,
-    inputContainerEl,
-    queueIndicatorEl,
-    inputWrapper,
-    inputEl,
-    sendButtonEl,
-    navRowEl,
-    contextRowEl,
-    eventCleanups: [],
-  };
 }
 
 export function updateSendButton(tab: TabData): void {
